Add files via upload
Nashville is the capital of the U.S. state of Tennessee and home to Vanderbilt University. Due to the increasing demand for houses in the locations, the price of houses is highly inflated. To get a clear picture of what is going on in Nashville I have downloaded public data from the government website and cleaned data for analysis purposes. Note: I have used this data to practice data cleaning, I will not work on the data visualization part, this is because as far as I am concerned this housing bubble is only temporary.